# backend/public/openapi.yaml
openapi: 3.0.3
info:
  title: Mobility Ops 360 API
  description: |
    タクシー業界向けAI配車システムAPI
    
    ## 主な機能
    - 🎙️ AI音声配車（Twilio統合）
    - 🚗 リアルタイムドライバーマッチング
    - 📍 位置情報トラッキング
    - 💰 動的料金計算
    - 🔐 セキュア認証
  version: 1.0.0
  contact:
    name: Mobility Ops 360 Support
    email: support@mobility360.jp
    url: https://mobility360.jp
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mobility360.jp
    description: Production server
  - url: https://mobi360-api.yukihamada.workers.dev
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /api/voice-dispatch/create:
    post:
      tags:
        - Voice Dispatch
      summary: AI音声配車リクエスト作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceDispatchRequest'
      responses:
        '201':
          description: 配車リクエスト作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceDispatchResponse'
  /api/voice-dispatch/twiml/{dispatchId}:
    post:
      tags:
        - Voice Dispatch
      summary: TwiML音声応答生成
      security: []
      parameters:
        - name: dispatchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TwiML生成成功
          content:
            application/xml:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoiceDispatchRequest:
      type: object
      required:
        - customerName
        - customerPhone
        - pickupLocation
        - destination
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        pickupLocation:
          type: string
        destination:
          type: string
        vehicleType:
          type: string
          enum: [standard, premium, wheelchair]
          default: standard
        notes:
          type: string
    VoiceDispatchResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            dispatchId:
              type: string
            assignedDriver:
              type: object
              properties:
                name:
                  type: string
                vehicleModel:
                  type: string
                vehiclePlate:
                  type: string
            estimatedArrival:
              type: integer
            twimlUrl:
              type: string
...